PirateFlow/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-dotenv==1.1.1
pydantic==2.11.7
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
aiosqlite==0.21.0
httpx==0.28.1
anthropic==0.52.0
